+++ apps/concord/src/main.ts
@@ -1,5 +1,6 @@
 import { ViteSSG } from "vite-ssg";
-// import "ternent-ui/styles";
+import "ternent-ui/styles.css";
+import "./style.css";
 import App from "./App.vue";
 import { routes } from "./route/index";