+++ apps/concord/src/route/playground/Playground.vue
@@ -1,1 +1,8 @@
-<template>playground</template>
+<script setup lang="ts"></script>
+<template>
+  <div class="h-screen w-screen flex flex-col">
+    <div class="flex-1 flex flex-col overflow-hidden">
+      <div class="flex-1"></div>
+    </div>
+  </div>
+</template>