+++ apps/concord/vite.config.ts
@@ -1,7 +1,12 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
+import tailwindcss from "@tailwindcss/vite";
 import Markdown from "vite-plugin-vue-markdown";
 
 export default defineConfig({
-  plugins: [vue({ include: [/\.vue$/, /\.md$/] }), Markdown()],
+  plugins: [
+    vue({ include: [/\.vue$/, /\.md$/] }),
+    tailwindcss(),
+    Markdown({ wrapperClasses: "prose" }),
+  ],
 });